flask_app/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
pymongo==4.8.0
pandas==2.2.2
numpy==1.24.4
python-dotenv==1.0.1

